# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Audiobookshelf-Hardcover Sync API
  description: |
    This API allows you to manage sync profiles for synchronizing audiobooks between Audiobookshelf and Hardcover.
    
    ## Authentication
    
    The API uses JWT authentication. Include the JWT token in the `Authorization` header:
    
    ```
    Authorization: Bearer <token>
    ```

    When authentication is enabled, newly created profiles belong to the
    authenticated user. For authenticated `/api/profiles` APIs, users and
    viewers can access only their owned profiles; viewers are read-only, while
    administrators can access all profiles. Profile API responses never echo
    `audiobookshelf_token` or `hardcover_token`, regardless of authentication
    or role. Legacy ownerless profiles are administrator-only: they are not
    listed or accessible to ordinary users, and there is currently no way to
    assign one of these profiles to a user. Foreign and nonexistent profile
    requests return a non-enumerating `404`. The public
    `/api/status` endpoint remains scalar and redacted. Authentication disabled
    preserves unrestricted profile access.
    
    ## Rate Limiting
    
    The API is rate limited to 100 requests per minute per IP address.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/drallgood/audiobookshelf-hardcover-sync
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.example.com/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Profile not found"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data
...
    Profile:
      type: object
      properties:
        id:
          type: string
          example: "profile-123"
        name:
          type: string
          example: "My Audiobooks"
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        last_successful_at:
          type: string
          format: date-time
          nullable: true
          description: Time the most recent non-dry-run run completed successfully.
...
    ProfileWithConfig:
      allOf:
        - $ref: '#/components/schemas/Profile'
        - type: object
          properties:
            config:
              $ref: '#/components/schemas/ProfileConfig'
...
    EditionCapability:
      type: object
      required: [can_create]
      properties:
        can_create:
          type: boolean
          description: True only after a definite successful scope probe, or for a dry-run profile where no mutation can occur.
        missing_scope:
          type: string
          description: The sanitized scope named by Hardcover when the probe received insufficient_scope.
          example: write:catalog:append
        reason:
          type: string
          enum: [unverified]
          description: The probe could not verify capability because of a transient or unclassified upstream result.
...
security:
  - BearerAuth: []
paths:
  /profiles:
    get:
      tags: [Profiles]
      summary: List all sync profiles
      responses:
        '200':
          description: List of sync profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags: [Profiles]
      summary: Create a new sync profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, audiobookshelfUrl, audiobookshelfToken, hardcoverToken]
              properties:
                id:
                  type: string
                  maxLength: 244
                  pattern: '^(?!\.\.?$)[A-Za-z0-9._~-]+$'
                  description: Up to 244 ASCII RFC 3986 unreserved characters; "." and ".." are not valid profile IDs.
                  example: "profile-123"
                name:
                  type: string
                  example: "My Audiobooks"
                audiobookshelfUrl:
                  type: string
                  example: "https://audiobooks.example.com"
                audiobookshelfToken:
                  type: string
                  example: "abs-token-123"
                hardcoverToken:
                  type: string
                  example: "hc-token-456"
                syncConfig:
                  $ref: '#/components/schemas/ProfileConfig/syncConfig'
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ProfileWithConfig'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profiles/{profileId}:
    get:
      tags: [Profiles]
      summary: Get a sync profile by ID
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
      responses:
        '200':
          description: Profile details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ProfileWithConfig'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags: [Profiles]
      summary: Update a sync profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Profile Name"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags: [Profiles]
      summary: Delete a sync profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
      responses:
        '200':
          description: Profile deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profiles/{profileId}/config:
    put:
      tags: [Profiles]
      summary: Update a profile's configuration
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                audiobookshelfUrl:
                  type: string
                  example: "https://new-audiobooks.example.com"
                audiobookshelfToken:
                  type: string
                  example: "new-abs-token-123"
                hardcoverToken:
                  type: string
                  example: "new-hc-token-456"
                syncConfig:
                  $ref: '#/components/schemas/ProfileConfig/syncConfig'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profiles/{profileId}/runs/{runId}/details:
    get:
      tags: [Status]
      summary: Get full details for an active or retained run of a profile
      description: Returns per-book outcomes when runId identifies the profile's active run or an exact retained terminal report. Unknown, other-profile, and evicted run IDs return 404; the endpoint never substitutes a newer run.
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
        - name: runId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123-1700000000000000000-1"
      responses:
        '200':
          description: Full active or retained run snapshot details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SyncSnapshot'
        '404':
          description: Profile, active run, or retained run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profiles/{profileId}/edition-capability:
    get:
      tags: [Editions]
      summary: Check whether a profile can create editions
      description: Uses the profile's Hardcover token to call insert_edition with an impossible book_id of -1 and a minimal body. Hardcover checks write:catalog:append before rejecting the book id, so this cannot create an edition. A validation error identifying the missing book means the scope is present; insufficient_scope returns can_create false and the missing scope. Any other response, including network errors, timeout, 401, 429 or 5xx, returns reason unverified and can_create false. A dry-run profile returns can_create true without a request because mutations are disabled. Definite results are cached for three minutes; unverified results for five seconds. Concurrent calls for the same profile and token share a probe. Token/config changes and profile deletion invalidate the cache. Requires profile write permission.
      parameters:
        - name: profileId
          in: path
          required: true
          schema: {type: string}
          example: profile-123
      responses:
        '200':
          description: Capability result, including an unverified result when Hardcover cannot be classified
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  data:
                    $ref: '#/components/schemas/EditionCapability'
        '400':
          description: Profile ID is missing
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '401':
          description: Authentication is enabled and the request is not authenticated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '403':
          description: Caller is a viewer without write permission
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '404':
          description: Profile not found or owned by another user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '503':
          description: Service is shutting down or the profile is being deleted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /profiles/{profileId}/runs/{runId}/books/{bookId}/edition-draft:
    get:
      tags: [Editions]
      summary: Get a draft Hardcover edition for a needs-review book
      description: Builds a draft of a new Hardcover edition from the Audiobookshelf item of a book that a run with available details (the active run or a retained one) marked needs_review and that has a Hardcover candidate (hardcover_book_id). Hyphens and spaces are removed from the ISBN, and when its checksum is valid the draft also fills the other ISBN form. The item must have an ASIN or a parseable ISBN, otherwise the request is rejected with 409. Requires write permission on the profile for authorization purposes. Previewing reads Audiobookshelf metadata and may call Audnex for an audiobook ASIN; it never constructs a Hardcover client or sends a Hardcover request, so no Hardcover token or scope is needed. The profile's Audiobookshelf token is used only on the server to fetch the item and is never returned.
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
        - name: runId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123-1700000000000000000-1"
        - name: bookId
          in: path
          required: true
          description: Audiobookshelf library item ID, from book_outcomes[].book_id of the run details.
          schema:
            type: string
          example: "li_abc123"
      responses:
        '200':
          description: Draft edition
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/EditionDraft'
        '401':
          description: Authentication is enabled and the request is not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Caller is a viewer without write permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile (including another user's profile), run, book record, or Audiobookshelf item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The book is not needs_review or has no numeric Hardcover book ID, its Audiobookshelf item has no ASIN or parseable ISBN, or the profile is being deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Audiobookshelf request failed; the message is generic and names only the service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: The service is shutting down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profiles/{profileId}/runs/{runId}/books/{bookId}/edition:
    post:
      tags: [Editions]
      summary: Create a Hardcover edition for a needs-review book
      description: Creates an edition for a needs_review book with a Hardcover candidate. The target book comes only from the run record. Only title, subtitle, ASIN, ISBN-10, ISBN-13, release date and edition information are editable; the server refetches Audiobookshelf and derives format, duration, language/country, authors, narrators and publisher. Author names must have an exact Hardcover match. Unmatched narrators and publishers are omitted with fixed warnings. Hardcover transport, GraphQL, rate-limit and timeout errors during resolution return sanitized 502 responses before insertion. A bad-checksum ISBN is accepted and reported with a fixed warning that does not imply an insert, since the request may be a dry run or reuse an existing edition. The profile token requires write:catalog:append; GET /profiles/{profileId}/edition-capability performs a safe scope probe. A create-time 403 insufficient_scope is returned with a fixed message and is not retried. The two-minute operation deadline starts before the ABS fetch and covers resolution, duplicate detection and insertion; caller disconnect does not cancel the accepted create. Service shutdown waits for in-flight creates until its configured shutdown timeout (default 2m30s; a shorter configured timeout can end the drain early). If dry-run is enabled, the request is validated and metadata may be read, but no mutation occurs and edition_id is 0. Matching editions of the same book are reused untouched; matches owned by another or unconfirmed book return 409. The unsupported cover-upload flow is not attempted. Requires profile write permission.
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
        - name: runId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123-1700000000000000000-1"
        - name: bookId
          in: path
          required: true
          description: Audiobookshelf library item ID, from book_outcomes[].book_id of the run details.
          schema:
            type: string
          example: "li_abc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditionCreateRequest'
      responses:
        '200':
          description: Edition created, already existing, or (in dry-run mode) not created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/EditionCreateResult'
        '400':
          description: Malformed or oversize body, an unknown field (including book_id and image_url), or a body that is not exactly one JSON object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication is enabled and the request is not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Caller is a viewer without write permission, or Hardcover returned insufficient_scope for write:catalog:append; create-time scope failures use a fixed error message and are not retried
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile (including another user's profile), run, book record, or Audiobookshelf item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The book is not needs_review or has no numeric Hardcover book ID, the profile is being deleted, its Audiobookshelf item has no ASIN or valid ISBN, an edition create for the same book is already in progress, or the submitted ASIN or ISBN belongs to an edition of another Hardcover book or one whose book could not be confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: The submitted edition fails validation or none of the exact Audiobookshelf author names matched a Hardcover author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Audiobookshelf or Hardcover request failed; the message is generic and names only the service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: The service is shutting down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /profiles/{profileId}/sync:
    post:
      tags: [Sync]
      summary: Start a sync for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
      responses:
        '202':
          description: Sync accepted and queued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AcceptedSyncRun'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Sync already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags: [Sync]
      summary: Cancel an in-progress sync for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
          example: "profile-123"
      responses:
        '200':
          description: Sync cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Profile not found or no sync in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'